pttrader/requirements.txt
pandas==1.1.5
python-telegram-bot==13.4.1
tinkoff-invest-openapi-client==0.0.7
yfinance==0.1.55